FIX: changes to maincomponents tree
