+++ interface/src/App.tsx
@@ -1,5 +1,5 @@
 import React, { useState } from "react"
-import Map from "./components/Map"
+import Map from "./components/Map/Map"
 
 const App: React.FC = () => {
   const [balance, setBalance] = useState(1000)
