+++ interface/src/components/Map/components/Pixel.tsx
