+++ interface/src/components/Map/components/heightmap.ts
