+++ interface/src/components/Static/Static.tsx
@@ -1,0 +1,8 @@
+import React, { useState } from "react"
+
+const Static: React.FC = () => {
+  return (
+    <></>
+  )
+}
+export default Static
